+++ frontend/src/domains/admin/Settings/Settings.tsx
@@ -5,24 +5,14 @@ import BellOutlineIcon from '@/components/icons/BellOutlineIcon';
 import OutOutlineIcon from '@/components/icons/OutOutlineIcon';
 import SendIcon from '@/components/icons/SendIcon';
 import TrashCanIcon from '@/components/icons/TrashCanIcon';
-import ConfirmModal from '@/components/ConfirmModal/ConfirmModal';
 import ProfileBox from './components/ProfileBox/ProfileBox';
 import SettingListBox from './components/SettingListBox/SettingListBox';
-import { useLogout } from './hooks/useLogout';
-import { useWithdraw } from './hooks/useWithdraw';
+import WithdrawModal from './components/WithdrawModal/WithdrawModal';
+import LogoutModal from './components/LogoutModal/LogoutModal';
 import { useNotificationSettingsPage } from './hooks/useNotificationSettingsPage';
-import {
-  settingsContainer,
-  withdrawModalContent,
-  withdrawWarningText,
-  withdrawDeletedItemsList,
-  withdrawAgreementLabel,
-  withdrawCheckbox,
-  withdrawSettingListBox,
-} from './Settings.style';
+import { settingsContainer, withdrawSettingListBox } from './Settings.style';
 import useAdminAuth from '@/domains/admin/Settings/hooks/useAdminAuth';
 import { usePWAPrompt } from '@/contexts/usePWAPrompt';
-import { useAppTheme } from '@/hooks/useAppTheme';
 
 declare global {
   interface Window {
@@ -37,8 +27,6 @@ type ModalState = { type: 'logout' } | { type: 'withdraw' } | { type: null };
 
 export default function Settings() {
   const [modalState, setModalState] = useState<ModalState>({ type: null });
-  const [isWithdrawChecked, setIsWithdrawChecked] = useState(false);
-  const theme = useAppTheme();
   const {
     isToggleEnabled,
     updateNotificationSetting,
@@ -47,13 +35,10 @@ export default function Settings() {
     needsPWAInstall,
   } = useNotificationSettingsPage();
   const { adminAuth, isLoading: isAdminAuthLoading } = useAdminAuth();
-  const { handleLogout } = useLogout();
-  const { handleWithdraw } = useWithdraw();
   const { showPrompt } = usePWAPrompt();
 
   const closeModal = () => {
     setModalState({ type: null });
-    setIsWithdrawChecked(false);
   };
 
   const handleToggleClick = () => {
@@ -132,42 +117,10 @@ export default function Settings() {
           />
         </div>
 
-        {modalState.type === 'logout' && (
-          <ConfirmModal
-            title='로그아웃'
-            message='정말 로그아웃하시겠습니까?'
-            onConfirm={handleLogout}
-            onClose={closeModal}
-          />
-        )}
+        {modalState.type === 'logout' && <LogoutModal onClose={closeModal} />}
 
         {modalState.type === 'withdraw' && (
-          <ConfirmModal
-            title='회원탈퇴'
-            onConfirm={handleWithdraw}
-            onClose={closeModal}
-            confirmText='탈퇴하기'
-            disabled={!isWithdrawChecked}
-          >
-            <div css={withdrawModalContent}>
-              <p css={withdrawWarningText(theme)}>
-                탈퇴 시 정보가 삭제되며 복구 불가합니다.
-              </p>
-              <ul css={withdrawDeletedItemsList(theme)}>
-                <li>관리자 계정 및 로그인 정보</li>
-                <li>서비스 내 모든 정보</li>
-              </ul>
-              <label css={withdrawAgreementLabel(theme)}>
-                <input
-                  type='checkbox'
-                  checked={isWithdrawChecked}
-                  onChange={(e) => setIsWithdrawChecked(e.target.checked)}
-                  css={withdrawCheckbox}
-                />
-                <span>위 내용을 모두 확인했고, 탈퇴하는 것에 동의합니다.</span>
-              </label>
-            </div>
-          </ConfirmModal>
+          <WithdrawModal onClose={closeModal} />
         )}
       </div>
     </>
